More invention action

+++ src/main/java/com/autonomouslogic/everef/model/api/ActivityCost.java
@@ -28,7 +28,7 @@ public class ActivityCost {
 	long productId;
 
 	@JsonProperty
-	long quantity;
+	double quantity;
 
 	@JsonProperty
 	@Schema(implementation = String.class)
